lab4: processor: Use memset() in sendRequest() and fix inconsistency
The bitwise AND with 0xFF is unnecessary when casting to uint8_t, so
remove it. Also use a memset here since it's equivalent to what this loop
is doing (and it does it faster).

Additionally, clean up the inconsistency with the arguments being passed to
cache->receiveRequest() (all of the arguments are members of Record r).

There's no need to set r.data to null when not writing either since it
defaults to null.

This commit does not alter the functionality of processor.cc; everything
still works as expected.