Update SRAM array with move constructor
There was a bug if you used emplace with the SRAMArray since it
incorrectly used the default copy constructor. The copy constructor
would copy the data pointer and it would be double-free'd when the
program exited. This fixes the error by implementing a *move*
constructor and deleting the copy constructor.

Using smart pointers may have also worked, but I couldn't figure it
out quickly.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>