Add written homework for week 7
I tried to make an informed guess on the topics since you deleted write
policies off the practice midterm. Remove anything you're not going to
cover, but I think you're going to cover this all.

I (tried to) based the access times in the last two questions off of
[this StackOverflow post](https://stackoverflow.com/questions/4087280/approximate-cost-to-access-various-caches-and-main-memory).
Feel free to correct it if it's wrong.