Fix Chisel links
Before, the chisel links lead to 404 errors as there are no local `chisel-notes`. I assume that the `chisel-notes` is intended to be the ones found in the dinocpu-sq20 repository: https://github.com/jlpteaching/dinocpu-sq20/tree/master/documentation/chisel-notes

Note that the "Building Systems" link does not work as the corresponding markdown file does not exist yet.